+++ src/dcos_e2e/backends/__init__.py
@@ -2,6 +2,7 @@
 Classes to allow backend-specific configuration for cluster backend types.
 """
 
+from ._aws import AWS
 from ._base_classes import ClusterBackend, ClusterManager
 from ._docker import Docker
 from ._existing_cluster import ExistingCluster as _ExistingCluster
@@ -9,6 +10,7 @@ from ._existing_cluster import ExistingCluster as _ExistingCluster
 __all__ = [
     'ClusterBackend',
     'ClusterManager',
+    'AWS',
     'Docker',
     '_ExistingCluster',
 ]