+++ src/dcos_e2e/distributions.py
@@ -14,3 +14,4 @@ class Distribution(Enum):
 
     CENTOS_7 = 1
     COREOS = 2
+    UBUNTU_16_04 = 3