+++ tests/test_cli/test_cli.py
@@ -144,7 +144,7 @@ class TestCreate:
               --docker-version [1.11.2|1.13.1|17.12.1-ce]
                                               The Docker version to install on the nodes.
                                               [default: 1.13.1]
-              --linux-distribution [centos-7|coreos]
+              --linux-distribution [centos-7|coreos|ubuntu-16.04]
                                               The Linux distribution to use on the nodes.
                                               [default: centos-7]
               --docker-storage-driver [aufs|overlay|overlay2]